% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{data}
\name{ecorys_colours}
\alias{ecorys_colours}
\title{List of Ecorys brand colours}
\format{
An object of class \code{list} of length 12.
}
\usage{
ecorys_colours
}
\value{
None
}
\description{
Use \code{\link{ecorys_colours}} to access Ecorys brand colours.
}
\examples{
ecorys_colours$blue

}
\keyword{datasets}
