% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{ecorys_mono_pal}
\alias{ecorys_mono_pal}
\title{Function to return monochrome blue palette of length \emph{n}}
\usage{
ecorys_mono_pal(n)
}
\value{
None
}
\description{
Use \code{\link{ecorys_mono_pal()}} to create a monochrome blue palette, where you can specify the number of shades.
}
\examples{
ecorys_mono_pal(4)

}
