% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{data}
\name{ecorys_palettes}
\alias{ecorys_palettes}
\title{List of palettes constructed from Ecorys brand colours}
\format{
An object of class \code{list} of length 4.
}
\usage{
ecorys_palettes
}
\description{
Use \code{\link{ecorys_palettes}} to access palettes of Ecorys brand colours.
}
\examples{
ecorys_palettes$primary

}
\keyword{datasets}
