% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_functions.R
\name{fct_case_when}
\alias{fct_case_when}
\title{fct_case_when()}
\usage{
fct_case_when(...)
}
\value{
None
}
\description{
\code{\link{fct_case_when}} extends \code{\link{dplyr::case_when}}. It outputs right hand side values as factors, where the order of resulting factor levels reflects the order in which arguments to the function are specified.
This function was developed by StackOverflow  user Dennis YL (see \href{https://stackoverflow.com/questions/49572416/r-convert-to-factor-with-order-of-levels-same-with-case-when}{here}).
}
\examples{
data \%>\% mutate(value = fct_case_when(value == 1 ~ "group_1",
value == 2 ~ "group_2"))

}
