% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_ecorys}
\alias{theme_ecorys}
\title{Theme to format ggplot2 charts in Ecorys style}
\usage{
theme_ecorys(x_grid = FALSE, y_grid = FALSE)
}
\arguments{
\item{x_grid}{Includes major grid lines for x axis}

\item{y_grid}{Includes major grid lines for y axis}
}
\value{
None
}
\description{
Use \code{\link{theme_ecorys}} to format ggplot2 plots in the Ecorys style.
}
\examples{
\dontrun{
tibble(x = seq(1,10,1), y = seq(1,10,1)) \%>\%
ggplot2::ggplot(aes(x = x, y = y)) +
geom_line() +
ggtitle("Plot in Ecorys theme") +
xlab("X axis") +
ylab("Y axis") +
theme_ecorys()
}

}
