% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_ecorys}
\alias{theme_ecorys}
\title{Theme to format ggplot2 charts in Ecorys style}
\usage{
theme_ecorys()
}
\value{
None
}
\description{
Use \code{\link{theme_ecorys}} to format ggplot2 plots in the Ecorys style.
}
\examples{
tibble(x = seq(1,10,1), y = seq(1,10,1)) \%>\%
ggplot(aes(x = x, y = y)) +
geom_line() +
theme_ecorys()

}
