% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{theme_ecorys}
\alias{theme_ecorys}
\title{Theme to format ggplot2 charts in Ecorys style}
\usage{
theme_ecorys(
  grid.major.x = FALSE,
  grid.major.y = FALSE,
  axis.line = FALSE,
  strip.background = TRUE
)
}
\arguments{
\item{grid.major.x}{Set to TRUE to display major grid lines for x axis (defaults to FALSE).}

\item{grid.major.y}{Set to TRUE to display major grid lines for y axis (defaults to FALSE).}

\item{axis.line}{Set to TRUE to display lines for x and y axis (defaults to FALSE),}

\item{strip.background}{Set to FALSE to hide facet label backgrounds (defaults to TRUE).}
}
\value{
None
}
\description{
Use \code{\link{theme_ecorys}} to format ggplot2 plots in the Ecorys style.
}
\examples{
example_data <- tibble(x = seq(1,10,1), y = seq(1,10,1))

ggplot(example_data = data, aes(x = x, y = y)) +
geom_line() +
theme_ecorys()


}
